requirements.txt
chromadb==0.4.23
ctransformers==0.2.27
InstructorEmbedding==1.0.1
langchain==0.1.9
langchain-community==0.0.22
llama-cpp-python==0.2.20
librosa==0.10.1
pyyaml==6.0.1
sentence-transformers==2.3.1
streamlit==1.31.1
streamlit-mic-recorder==0.0.4
transformers==4.38.1